requirements.txt
django
 djangorestframework 

 